changes float32

+++ module1/budget_1.go
@@ -2,7 +2,7 @@ package module1
 
 // Budget stores budget information
 type Budget struct {
-	Max   float64
+	Max   float32
 	Items []Item
 }
